requirements.txt
adal==1.2.4
applicationinsights==0.11.9
argcomplete==1.12.0
astroid==2.3.3
azure-cli-core==2.11.1
azure-cli-telemetry==1.0.6
azure-common==1.1.25
azure-core==1.8.1
azure-identity==1.3.1
azure-mgmt-core==1.2.0
azure-mgmt-resource==10.2.0
azure-mgmt-resourcegraph==2.0.0
azure-storage-blob==12.3.1
azure-storage-file-datalake==12.0.1
bcrypt==3.2.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
humanfriendly==8.2
idna==2.9
importlib-metadata==1.7.0
isodate==0.6.0
isort==4.3.21
jmespath==0.10.0
knack==0.7.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
msal==1.0.0
msal-extensions==0.1.3
msgraph==0.1.2
msrest==0.6.13
msrestazure==0.6.4
oauthlib==3.1.0
paramiko==2.7.2
pep8==1.7.1
pkginfo==1.5.0.1
portalocker==1.7.0
pycparser==2.20
Pygments==2.7.1
PyJWT==1.7.1
pylint==2.4.4
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyreadline==2.1
python-dateutil==2.8.1
pywin32==227
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
tabulate==0.8.7
typed-ast==1.4.1
urllib3==1.25.9
virtualenv==16.5.0
wrapt==1.11.2
zipp==3.1.0
